Refactor coarse alignment gui into a single module
